rag-pipeline/docker/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.55.0
qdrant-client>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.10.0
